все сломал

+++ src/main/java/com/example/springlesson3/service/ProductServiceImp.java
@@ -59,12 +59,12 @@ public class ProductServiceImp implements ProductService {
     }
 
     @Override
-    public Product getProductById(int id) {
+    public Product getProductById(Long id) {
         return productRepository.getById(id);
     }
 
     @Override
-    public void deleteProduct(int id) {
+    public void deleteProduct(Long id) {
         productRepository.deleteById(id);
     }
